refactor: extract activity components
